kernel: Fix 'mult' overflow if using 10 Hz
Change 'mult' to 64-bit (u64) to fix overflow warning if using 10 Hz jiffies.

./include/vdso/jiffies.h:9:19: warning: unsigned conversion from 'long int' to 'unsigned int' changes value from '6400000000' to '2105032704' [-Woverflow]
    9 | #define TICK_NSEC ((NSEC_PER_SEC+HZ/2)/HZ)
      |                   ^
kernel/time/jiffies.c:38:35: note: in expansion of macro 'TICK_NSEC'
   38 |         .mult                   = TICK_NSEC << JIFFIES_SHIFT, /* details above */
      |                                   ^~~~~~~~~

+++ kernel/time/clocksource.c
@@ -54,7 +54,7 @@ static noinline u64 cycles_to_nsec_safe(struct clocksource *cs, u64 start, u64 e
  * factors.
  */
 void
-clocks_calc_mult_shift(u32 *mult, u32 *shift, u32 from, u32 to, u32 maxsec)
+clocks_calc_mult_shift(u64 *mult, u32 *shift, u32 from, u32 to, u32 maxsec)
 {
 	u64 tmp;
 	u32 sft, sftacc= 32;
